requirements.txt
django>=2.0.0
django-filter>=2.0.0
django-model-utils>=3.1.2
django-widget-tweaks>=1.4.2
djangorestframework>=3.9.0

channels>=2.1.3
channels_redis>=2.3.0

